Reference

Full description of the functions implemented in PolieDRO.jl.

Contents

Reference

Contents

Index

Index

PolieDRO.HullsInfo

PolieDRO.PolieDROModel

PolieDRO.build_model

PolieDRO.build_model

PolieDRO.build_model

PolieDRO.calculate_convex_hulls

PolieDRO.calculate_hulls_probabilities!

PolieDRO.solve_model!

PolieDRO.HullsInfo — Type

Convex hulls and probabilities structure

Stores information about the convex hulls that can be reused for separate PolieDRO models being trained in the same dataset Not having to recalculate convex hulls drastically reduces model training time

Fields

index_sets::Vector{Vector{Int64}}: List of indices of each hull in the form:

    [
        [
            (index of point 1 in outermost hull),
            (index of point 2 in outermost hull),
            etc.
        ],
        [
            (index of point 1 in inner hull 1),
            (index of point 2 in inner hull 1),
            etc.
        ],
        (one entry per convex hull)
    ]

non_vertices::Vector{Int64}: List of indices which are not vertices

The last hull includes its inner points, so some of the points inside it are not vertices of the hull

significance_level::Float64: The significance level used to calculate the hulls' associated probabilities

probabilities::Vector{Vector{Int64}}: The confidence interval associated with each convex hull in the form:

    [
        [
            (lower end of hull 1 interval),
            (upper end of hull 1 interval)
        ],
        [
            (lower end of hull 2 interval),
            (upper end of hull 2 interval)
        ],
        (one entry per convex hull)
    ]

PolieDRO.PolieDROModel — Type

PolieDRO model structure

Stores PolieDRO model information, such as optimized parameters and the inner JuMP model

Fields

model::GenericModel: JuMP model where the DRO problem is defined

β0::Float64: The intercept term of the model solution

β1::Vector{Float64}: The vector of coefficients of the model solution

optimized::Bool: If the model has been solved or not

PolieDRO.build_model — Method

build_model(X, y, loss_function, point_evaluator; hulls=nothing, significance_level=0.05)

Build model function (custom loss function version)

Calculates the convex hulls and probabilities associated with the given data and builds the PolieDRO model for the specified loss function.

Arguments

X::Matrix{Float64}: Matrix N x D of points in which the model is trained (N = number of points, D = dimension of points)

y::Vector{Float64}: Dependent variable vector relative to the points in the matrix X (size N)

loss_function::Function: A loss function to be used in the PolieDRO formulation

Has to be convex! (This is not checked)

This function defines the solver you will be able to use (for instance, if you use a nonlinear function you will need a nonlinear solver)

Function must have a method f(x::Vector{T}, y::T, β0::T, β1::Vector{T}) where T is Float64

point_evaluator::Function: A function to evaluate a given point x and the optimized parameters β0, β1

Function must have a method f(x::Vector{T}, β0::T, β1::Vector{T}) where T is Float64

Optionals

hulls::Union{HullsInfo,Nothing}: Pre-calculated convex hulls, output of calculate_convex_hulls.

Default value: nothing (calculates hulls)

Pre-calculated hulls are useful if many models are being tested on the same data, since the hulls only have to be calculated once.

significance_level::Float64: Used to define a confidence interval for the probabilities associated to the hulls

Default value: 0.05

silent::Bool: Sets the flag to build the hulls silently (without logs)

Default value: true

Returns

An unsolved PolieDROModel struct, that can be solved using the solve_model function

A predictor function, which takes the solved model and a matrix of points X and predicts their y

Assertions

X and y must match in sizes (N x D and N)

N must be larger than D (no high dimensional problems)

No Infinite or NaN values in either X or y

No duplicate points in X

PolieDRO.build_model — Method

build_model(X, y, loss_function, point_evaluator; hulls=nothing, significance_level=0.05)

Build model function (pre-implemented loss functions)

Calculates the convex hulls and probabilities associated with the given data and builds the PolieDRO model for a pre-implemented loss function.

Arguments

X::Matrix{Float64}: Matrix N x D of points in which the model is trained (N = number of points, D = dimension of points)

y::Vector{Float64}: Dependent variable vector relative to the points in the matrix X (size N)

loss_function::LossFunctions: One of the given loss functions implemented in the enumerator

significance_level::Float64: Used to define a confidence interval for the probabilities associated to the hulls

Default value: 0.05

silent::Bool: Sets the flag to build the hulls silently (without logs)

Default value: true

Returns

An unsolved PolieDROModel struct, that can be solved using the solve_model function

A predictor function, which takes the solved model and a matrix of points X and predicts their y

Assertions

X and y must match in sizes (N x D and N)

N must be larger than D (no high dimensional problems)

No Infinite or NaN values in either X or y

No duplicate points in X

For classification models (hinge and logistic loss) all values in y must be either 1 or -1

PolieDRO.build_model — Method

build_model(X, y, loss_functions, point_evaluator; hulls=nothing, significance_level=0.05, silent=true)

Build model function (custom epigraph version)

Calculates the convex hulls and probabilities associated with the given data and builds the PolieDRO model for the specified loss function. The loss function in this case is a maximum of a group of functions, modeled as an epigraph. This is used, for instance, in the hinge loss function.

Arguments

X::Matrix{Float64}: Matrix NxD of points in which the model is trained (N = number of points, D = dimension of points)

y::Vector{Float64}: Dependent variable vector relative to the points in the matrix X (size N)

loss_function::Vector{Function}: A list of functions to be used in the PolieDRO formulation, the loss function will be an epigraph above all those

They have to be convex! (This is not checked)

These functions defines the solver you will be able to use (for instance, if you use a nonlinear function you will need a nonlinear solver)

Functions must have a method f(x::Vector{T}, y::T, β0::VariableRef, β1::Vector{VariableRef}) where T is Float64

This method allows you to use multiple linear functions instead of a piecewise linear one and use a linear solver

point_evaluator::Function: A function to evaluate a given point x and the optimized parameters β0, β1

Function must have a method f(x::Vector{T}, β0::T, β1::Vector{T}) where T is Float64

Optionals

hulls::Union{HullsInfo,Nothing}: Pre-calculated convex hulls, output of calculate_convex_hulls.

Default value: nothing (calculates hulls)

Pre-calculated hulls are useful if many models are being tested on the same data, since the hulls only have to be calculated once.

significance_level::Float64: Used to define a confidence interval for the probabilities associated to the hulls

Default value: 0.05

silent::Bool: Sets the flag to build the hulls silently (without logs)

Default value: true

Returns

An unsolved PolieDROModel struct, that can be solved using the solve_model function

A predictor function, which takes the solved model and a matrix of points X and predicts their y

Assertions

X and y must match in sizes (N x D and N)

N must be larger than D (no high dimensional problems)

No Infinite or NaN values in either X or y

No duplicate points in X

PolieDRO.calculate_convex_hulls — Method

calculate_convex_hulls(X; silent=true)

Calculates the convex hulls associated with the given matrix of points

Calculates the outer hull, removes the points present in the hulls, calculate the inner hull, repeat to calculate all hulls *Currently does this with a LP model where it tries to create each point with a convex combination of the rest *Looking into faster ways of solving this problem, since it is by far the biggest bottleneck

Arguments

X::Matrix{Float64}: Matrix NxD of points to calculate the hulls (N = number of points, D = dimension of points)

silent::Bool: Sets the flag to build the hulls silently (without logs)

Default value: true

Returns

Convex hulls structure that can be passed to a model builder function

PolieDRO.calculate_hulls_probabilities! — Method

calculate_hulls_probabilities!(hulls_struct, significance_level)

Calculates the convex hulls and probabilities associated with the given data and builds the PolieDRO model for the specified loss function. Stores this data in the hulls struct.

Arguments

hulls_struct::HullsInfo: Structure of convex hulls as returned by the calculateconvexhulls function

significance_level::Float64: Used to define a confidence interval for the probabilities associated to the hulls (read more in the README.md)

Data stored in the struct:

List of tuples of the probability intervals associated with each convex hull in the form: [ [(lower end of the interval), (upper end of the interval),], <= interval for outermost hull [(lower end of the interval), (upper end of the interval),], <= interval for inner hull 1 etc. ] List is the size of the given list of hulls. First tuple is always (1,1) because first hull contains all points

Assertions

Significance_level must be positive within 0 and 1

PolieDRO.solve_model! — Method

solve_model!(model; optimizer=Ipopt.Optimizer, silent=true, attributes=nothing)

Solve model function

Uses the given solver to solve the PolieDRO model. Modifies the struct with the results and sets the optimized bool in it to true

Arguments

model::PolieDROModel: A PolieDRO model struct, as given by the build_model function, to be solved

optimizer: An optimizer as the ones used to solve JuMP models

Default: Ipopt optimizer

NOTE: For the logistic and MSE models, a nonlinear solver is necessary

silent::Bool: Sets the flag to solve the model silently (without logs)

Default value: true

attributes::Union{Nothing, Dict{String}}: Sets optimizer attribute flags given a dictionary with entries attribute => value

Default value: nothing